# Optimize purge for oci manifests, indexes and docker manifest list scenarios through improved concurrency