--- a/cmd/acr/annotate.go
+++ b/cmd/acr/annotate.go
@@ -193,7 +193,7 @@ func annotateTags(ctx context.Context,
 		}
 		lastTag = newLastTag
 		if manifestsToAnnotate != nil {
-			count := len(*manifestsToAnnotate)
+			count := len(manifestsToAnnotate)
 			if !dryRun {
 				_, annotateErr := annotator.Annotate(ctx, manifestsToAnnotate)
 				if annotateErr != nil {
@@ -219,7 +219,7 @@ func getManifestsToAnnotate(ctx context.Context,
 	loginURL string,
 	repoName string,
 	filter *regexp2.Regexp,
-	lastTag string, artifactType string, dryRun bool) (*[]string, string, error) {
+	lastTag string, artifactType string, dryRun bool) ([]string, string, error) {
 
 	resultTags, err := acrClient.GetAcrTags(ctx, repoName, "timedesc", lastTag)
 	if err != nil {
@@ -264,7 +264,7 @@ func getManifestsToAnnotate(ctx context.Context,
 		}
 
 		newLastTag = common.GetLastTagFromResponse(resultTags)
-		return &manifestsToAnnotate, newLastTag, nil
+		return manifestsToAnnotate, newLastTag, nil
 	}
 	return nil, "", nil
 }
@@ -306,7 +306,7 @@ func annotateUntaggedManifests(ctx context.Context,
 		}
 		annotatedManifestsCount += manifestsCount
 	} else {
-		annotatedManifestsCount = len(*manifestsToAnnotate)
+		annotatedManifestsCount = len(manifestsToAnnotate)
 	}
 
 	return annotatedManifestsCount, nil