--- a/cmd/acr/annotate.go
+++ b/cmd/acr/annotate.go
@@ -287,7 +287,7 @@ func annotateUntaggedManifests(ctx context.Context,
 	// Contrary to getTagsToAnnotate, getManifests gets all the manifests at once.
 	// This was done because if there is a manifest that has no tag but is referenced by a multiarch manifest that has tags then it
 	// should not be annotated.
-	manifestsToAnnotate, err := common.GetUntaggedManifests(ctx, acrClient, loginURL, repoName, true, nil, dryRun)
+	manifestsToAnnotate, err := common.GetUntaggedManifests(ctx, poolSize, acrClient, loginURL, repoName, true, nil, dryRun)
 	if err != nil {
 		return -1, err
 	}