--- a/go.mod
+++ b/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/Azure/go-autorest/autorest v0.11.30
 	github.com/Azure/go-autorest/autorest/adal v0.9.24
 	github.com/Azure/go-autorest/tracing v0.6.1
-	github.com/alitto/pond/v2 v2.3.4
+	github.com/alitto/pond/v2 v2.4.0
 	github.com/dlclark/regexp2 v1.11.5
 	github.com/docker/cli v28.2.2+incompatible
 	github.com/golang-jwt/jwt/v4 v4.5.2