**Purpose of the PR**
The primary goal of this PR is to optimize the purge process by introducing a new pool mechanism to the fetch portion of purge which today is one of the main limiting factors.  In a registry with a lot of OCI artifacts or indexes the current purge requires a manifest fetch for every single one. If these are done sequentially the probability of this taking too long is quite high. Instead, we introduce a pool mechanism that allows us to fetch multiple manifests concurrently. This is meant to address concerns like those brought up in: #457 

***Changelog***
- Optimized fetch path by adding concurrency to manifest fetching
- Properly support nested indexes (Nested indexes should not get cleaned up if their parent is still in use)
- Added support for OCI index referrers
- Refactored the purge flow for increased readability. 
- Removed the separate dry run scenario and integrated this into the code 
- Changed existing pool mechanisms to use pond https://github.com/alitto/pond

***Validations***
- Unit tests have been added to account for new scenarios above, other unit tests have been updated to account for the change.
- Ran validations against real registries with the following configurations with a before and after time:

## Performance Improvements to Scanning
Most of the changes in this PR are optimizing the scanning of repositories, specifically for OCI index and OCI artifact scenarios as well as those involving child manifests. I ran some validation to fully cleanup a mix of distinct repos with 5000 untagged manifests and 200 untagged manifests each using dry-run (to avoid actually deleting) and --concurrency of 16 and observed the following performance improvements:

| Manifest Mix   | Old Binary (s) | New Binary (s) | Improvement |
|----------------|----------------|----------------|----------|
| 100% Docker | 58 | 55 | 5.2% |
| 100% OCI | 737 | 68 | 90.8% |
| 100% Lists | 54| 54 | 0.0% |
| 100% Indexes | 433 | 73 | 83.1% |
| Equal Split | 730 | 73 | 90.0% |

As can be observed there is no major change for registries using just docker manifests, however when we are dealing with a mix of manifest types or OCI indexes and OCI manifests the performance improvement is pretty substantial.

## Performance Comparison Matrixes When Actually carrying out deletes
The fixes in this PR are only substantial for situations that require cleanup of larger repositories. Purge itself seems to be rather hard to time as the actual delete times can have a large influence on the result but most of the changes are to the acquisition pathway. Nonetheless, to evaluate if there is otherwise any performance degradation, I ran the following scenarios and there seems to still be an overall improvement though less marked than when just accounting for acquisition, it might be worthwhile to bring some integration tests to do validation like this in the future but that's out of the scope of this PR.

Based on the binary execution log, here's a comprehensive summary of the ACR CLI purge performance comparison across many scenarios:

### ACR CLI Purge Performance Comparison Results

### Test Summary
- **Ran On**:	AMD EPYC 7763 64-Core Processor, 2445 Mhz, 8 Core(s), 16 Logical Processor(s) with 64GB RAM
- **Registry**: esreyconversion.azurecr.io, Premium SKU
- **Test Period**: July 1-2, 2025

### Performance Results by Scale

#### Small Scale (50 Tags + 100 Untagged)
| Manifest Mix | Old Binary | New Binary | Improvement | Manifests Deleted |
|--------------|------------|------------|-------------|-------------------|
| 100% Docker | 39s | 36s | **7.7%** | 170 → 170 |
| 100% OCI | 48s | 31s | **35.4%** | 170 → 170 |
| 100% Lists | 42s | 41s | **2.4%** | 170 → 170 |
| 100% Indexes | 47s | 39s | **17.0%** | 170 → 170 |
| Balanced (40/30/20/10) | 46s | 40s | **13.0%** | 170 → 170 |
| OCI Heavy (10/70/10/10) | 45s | 25s | **44.4%** | 170 → 170 |
| Equal Split (25/25/25/25) | 53s | 42s | **20.8%** | 170 → 171 |

#### Medium Scale (200 Tags + 400 Untagged)
| Manifest Mix | Old Binary | New Binary | Improvement | Manifests Deleted |
|--------------|------------|------------|-------------|-------------------|
| 100% Docker | 111s | 105s | **5.4%** | 620 → 621 |
| 100% OCI | 121s | 112s | **7.4%** | 620 → 620 |
| 100% Lists | 106s | 105s | **0.9%** | 624 → 624 |
| 100% Indexes | 151s | 94s | **37.7%** | 623 → 624 |
| Balanced | 115s | 101s | **12.2%** | 620 → 621 |
| OCI Heavy | 141s | 106s | **24.8%** | 620 → 620 |
| Equal Split | 140s | 104s | **25.7%** | 621 → 622 |

#### Large Scale (300 Tags + 1,000 Untagged)
| Manifest Mix | Old Binary | New Binary | Improvement | Manifests Deleted |
|--------------|------------|------------|-------------|-------------------|
| 100% Docker | 175s | 162s | **7.4%** | 1,320 → 1,320 |
| 100% OCI | 245s | 170s | **30.6%** | 1,320 → 1,320 |
| 100% Lists | 183s | 166s | **9.3%** | 1,322 → 1,327 |
| 100% Indexes | 234s | 192s | **17.9%** | 1,324 → 1,325 |
| Balanced | 182s | 191s | **-4.9%** | 1,320 → 1,320 |
| OCI Heavy | 223s | 10s* | ⚠️ | 1,321 → error |
| Equal Split | 356s | 170s | ⚠️ | 2,587 → 1,320 |

*Authentication error occurred

#### Very Large Scale (300 Tags + 10,000 Untagged)
| Manifest Mix | Old Binary | New Binary | Improvement | Manifests Deleted |
|--------------|------------|------------|-------------|-------------------|
| 100% Docker | 881s | 883s | **-0.2%** | 10,320 → 10,320 |
| 100% OCI | 1,565s | 939s | **40.0%** | 10,320 → 10,321 |
| 100% Lists | 885s | 839s | **5.2%** | 10,338 → 10,346 |
| 100% Indexes | 1,548s | 854s | **44.8%** | 10,342 → 10,351 |
| Balanced | 1,119s | 854s* | **23.7%*** | 10,326 → 10,279 |
| OCI Heavy | 87028s | 961s | ***98.9%*** | 10,328 → 10,325 |
| Equal Split | 1267s | 904s | ***28.7%*** | 10,330 → 10,325 |

⚠️️ (needs to be retried if errored or otherwise have to complete it)

### Key Findings
1. **Consistent improvements for OCI-heavy workloads**: 30-44% faster. There was also a massive outlier leading to an almost 99% reduction, but I am not entirely sure why the original took so long for that case so Ill consider it anomalous for now.
2. **Consistent results for Docker-heavy workloads**

The optimization appears successful in reducing manifest fetching overhead and improving concurrent processing efficiency.